+++ lib/main.dart
@@ -1,5 +1,5 @@
+import 'package:education/ui/landing_page/landing_page.dart';
 import 'package:flutter/material.dart';
-import 'Landing.dart';
 
 void main() {
   runApp(MyApp());
@@ -16,6 +16,6 @@ class MyApp extends StatelessWidget {
           primarySwatch: Colors.blue,
           visualDensity: VisualDensity.adaptivePlatformDensity,
         ),
-        home: Landing());
+        home: LandingPage());
   }
 }
